Assign proper objects to Subscription
